adk_agents_project/requirements.txt

google-adk>=0.5.0
google-genai>=0.2.0
